+++ docs/_code/prepare_climate.py
@@ -35,6 +35,10 @@ tasks.catchment_area(gdir)
 tasks.catchment_width_geom(gdir)
 tasks.catchment_width_correction(gdir)
 cfg.PARAMS['baseline_climate'] = 'HISTALP'
+cfg.PARAMS['use_tstar_calibration'] = True
+cfg.PARAMS['use_winter_prcp_factor'] = False
+cfg.PARAMS['hydro_month_nh'] = 10
+cfg.PARAMS['hydro_month_sh'] = 4
 tasks.process_histalp_data(gdir)
 with warnings.catch_warnings():
     warnings.filterwarnings("ignore", category=FutureWarning)